Simplify uv CI task orchestration
