+++ .github/workflows/CI.yml
@@ -37,11 +37,20 @@ jobs:
     steps:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
 
+      - uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
+        id: python
+        with:
+          python-version: ${{ matrix.python-version }}
+          allow-prereleases: true
+
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
-          enable-cache: true
-          cache-dependency-glob: "uv.lock"
-          python-version: ${{ matrix.python-version }}
+          version: "0.11.19"
+          enable-cache: false
+
+      - name: Pin uv to the build interpreter
+        shell: bash
+        run: echo "UV_PYTHON=${{ steps.python.outputs.python-path }}" >> "$GITHUB_ENV"
 
       - uses: extractions/setup-just@53165ef7e734c5c07cb06b3c8e7b647c5aa16db3 # v4
 
@@ -55,7 +64,6 @@ jobs:
         with:
           manylinux: auto
           target: ${{ matrix.target }}
-          rust-toolchain: stable
           sccache: "true"
           args: >
             --release
@@ -85,7 +93,6 @@ jobs:
         uses: PyO3/maturin-action@e83996d129638aa358a18fbd1dfb82f0b0fb5d3b # v1
         with:
           manylinux: auto
-          rust-toolchain: stable
           command: build
           target: ${{ matrix.target }}
           sccache: "true"
@@ -151,6 +158,7 @@ jobs:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
+          version: "0.11.19"
           enable-cache: true
           cache-dependency-glob: "uv.lock"
           python-version: ${{ matrix.python-version }}
@@ -181,6 +189,7 @@ jobs:
           fetch-depth: 0
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
+          version: "0.11.19"
           enable-cache: true
           cache-dependency-glob: "uv.lock"
           python-version: "3.13"
@@ -221,6 +230,7 @@ jobs:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
+          version: "0.11.19"
           enable-cache: true
           cache-dependency-glob: "uv.lock"
           python-version: "3.13"
@@ -253,6 +263,7 @@ jobs:
           args: --release --sdist -o wheels
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
+          version: "0.11.19"
           enable-cache: true
           cache-dependency-glob: "uv.lock"
       - uses: extractions/setup-just@53165ef7e734c5c07cb06b3c8e7b647c5aa16db3 # v4
@@ -273,6 +284,7 @@ jobs:
       - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
       - uses: astral-sh/setup-uv@08807647e7069bb48b6ef5acd8ec9567f424441b # v8.1.0
         with:
+          version: "0.11.19"
           enable-cache: true
           cache-dependency-glob: "uv.lock"
           python-version: "3.13"
@@ -283,13 +295,13 @@ jobs:
           merge-multiple: true
           path: wheels
       - name: install dependencies
-        run: just ci-bench-codspeed-setup
+        run: just ci-bench-codespeed-setup
       - name: Run benchmarks
         uses: CodSpeedHQ/action@76578c2a7ddd928664caa737f0e962e3085d4e7c # v3
         with:
           token: ${{ secrets.CODSPEED_TOKEN }}
           mode: ${{ matrix.mode }}
-          run: just bench-codspeed-run
+          run: just bench-codespeed-run
 
   fmt:
     name: rustfmt
