+++ python/serpyco_rs/_impl.pyi
@@ -1,23 +1,4 @@
-from typing import Any, Generic, TypeVar
-
-from ._type_info import BaseType as _BaseType
-
-_T = TypeVar('_T')
-
-class ValidationError(Exception):
-    message: str
-
-class ErrorItem:
-    message: str
-    instance_path: str
-
-    def __init__(self, message: str, instance_path: str) -> None: ...
-
-class SchemaValidationError(ValidationError):
-    errors: list[ErrorItem]
-
-class Serializer(Generic[_T]):
-    def __init__(self, py_class: _BaseType, naive_datetime_to_utc: bool, max_recursion_depth: int) -> None: ...
-    def dump(self, value: _T) -> Any: ...
-    def load(self, data: Any) -> _T: ...
-    def load_query_params(self, data: Any) -> _T: ...
+from ._serpyco_rs import ErrorItem as ErrorItem
+from ._serpyco_rs import SchemaValidationError as SchemaValidationError
+from ._serpyco_rs import Serializer as Serializer
+from ._serpyco_rs import ValidationError as ValidationError
